countries_api-main/requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==4.2.1
djangorestframework==3.14.0
mysqlclient==2.1.1
pytz==2023.3
sqlparse==0.4.4
typing-extensions==4.7.1
